requirements.txt
Flask==2.2.2
opencv-python==4.7.0.72
Pillow==9.5.0
numpy==1.24.3
requests==2.31.0
